requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data processing
h5py>=3.7.0
tables>=3.8.0

# Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration
PyYAML>=6.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.64.0
wandb>=0.13.0
tensorboard>=2.10.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991